refactor(ui): Reorganized frontend component folder structure into subfolder where needed. Included also the ability to delete a single pipeline.
