+++ frontend/src/components/Pipelines/Source.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Typography, List, ListItem, ListItemText, Box } from "@mui/material";
-import { PipeLine } from "../types/types";
+import { PipeLine } from "../../types/types";
 
 const Source = ({ pipeLineData }: { pipeLineData: PipeLine }) => {
   return (
